LayoutTests/editing/deleting/2610675-1.html

<html>
<head>
<style>
blockquote[type="cite"] {
    color: blue;
    margin: 0px;
    padding-left: 4px;
    border-left: 2px solid blue;
}
</style>
</head>
<body>
<div id="description">This tests to see that if the caret is in an empty quoted paragraph, that paragraph is in content that preserve newlines, and there's no quoted content before that paragraph, then pressing delete removes that paragraph's quoting.  Below you should see the attribution line, an empty paragraph, an unquoted paragraph, and then a quoted paragraph with "quoted" in it.</div>
<div id="edit" contentEditable="true">
<div><br></div>
<div>On Tuesday, Justin wrote:</div>
<div><br></div>
<blockquote type="cite" id="blockquote"><div style="white-space: pre;">
quoted</div></blockquote>
</div>

<script>
if (window.layoutTestController)
    window.layoutTestController.dumpAsText();
blockquote = document.getElementById("blockquote");
window.getSelection().setPosition(blockquote, 0);
document.execCommand("Delete");
document.execCommand("InsertText", false, "not quoted");
if (window.layoutTestController)
    document.body.innerText = document.getElementById("description").innerText + "\n" + document.getElementById("edit").innerHTML;
</script>

</body>
</html>
