third_party/blink/web_tests/editing/deleting/2610675-1-expected.txt

This tests to see that if the caret is in an empty quoted paragraph, that paragraph is in content that preserve newlines, and there's no quoted content before that paragraph, then pressing delete removes that paragraph's quoting and removes the empty line. Below you should see the attribution line, an unquoted paragraph, and then a quoted paragraph with "quoted" in it.

<div><br></div>
<div>On Tuesday, Justin wrote:</div>
<div>not quoted</div><blockquote type="cite" id="blockquote"><div style="white-space: pre;">quoted</div></blockquote>
